Updated naming conventions, added Command Phase
