+++ README.md
@@ -1,1 +1,10 @@
-# WhirhammurApp
+# Whirhammur App
+
+An Android app for the combat process in a certain table top roleplaying game.  in **StackEdit**. If you want to learn about StackEdit, you can read me. If you want to play with Markdown, you can edit me. Once you have finished with me, you can create new files by opening the **file explorer** on the left corner of the navigation bar.
+
+
+## Things to implement
+
+- Basic functionality of going through the  combat process with hardcoded modifiers for my army in particular
+-  A button to switch between attacking and defense
+- The ability to switch armies and get different modifiers
