+++ app/src/main/res/values/strings.xml
@@ -1,11 +1,25 @@
 <resources>
-    <string name="app_name">Test2</string>
+    <string name="app_name">Whirhammur Battle App</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">First Fragment</string>
-    <string name="second_fragment_label">Second Fragment</string>
-    <string name="next">Next</string>
-    <string name="previous">Previous</string>
+    <string name="start_fragment_label">Start Fragment</string>
+    <string name="battleshock_fragment_label">Battle Shock Fragment</string>
 
-    <string name="lorem_ipsum">Counting App</string>
+    <!-- Start Fragment -->
+    <string name="title_command">1. Command Phase</string>
+    <string name="details_command">
+        - Both players gain 1CP \n
+        - Resolve additional rules \n
+        - Use stratagems \n
+    </string>
+    <string name="title_battleshock">2. Battle Shock</string>
+    <string name="details_battleshock">
+        Are any units below half strength? \n
+        (Unit with starting strength of 1 has wounds less than half or number of models less than half starting strength) \n
+        No - proceed to MOVEMENT \n
+        Yes - continue to test \n
+    </string>
+    <string name="test_battleshock_button_text">Test BATTLESHOCK</string>
+    <string name="goto_movement_button_text">Go To MOVEMENT</string>
+    <!-- TODO: have a faction selection on startup to create additional rules -->
 </resources>
